# Accelerate node lookup in presence of many nodes without proxies