--- a/nestkernel/node_manager.cpp
+++ b/nestkernel/node_manager.cpp
@@ -524,13 +524,14 @@ NodeManager::ensure_valid_thread_local_ids()
       {
         wfr_nodes_vec_[ tid ].clear();
 
-        const size_t num_thread_local_wfr_nodes = std::count_if(local_nodes_[tid].begin(), local_nodes_[tid].end(),
-        	                                                	[](const SparseNodeArray::NodeEntry& elem) { return elem.get_node()->node_uses_wfr_; });
+        const size_t num_thread_local_wfr_nodes = std::count_if( local_nodes_[ tid ].begin(),
+          local_nodes_[ tid ].end(),
+          []( const SparseNodeArray::NodeEntry& elem ) { return elem.get_node()->node_uses_wfr_; } );
         wfr_nodes_vec_[ tid ].reserve( num_thread_local_wfr_nodes );
 
-        auto node_it = local_nodes_[tid].begin();
+        auto node_it = local_nodes_[ tid ].begin();
         size_t idx = 0;
-        for (  ; node_it < local_nodes_[tid].end() ; ++node_it, ++idx )
+        for ( ; node_it < local_nodes_[ tid ].end(); ++node_it, ++idx )
         {
           auto node = node_it->get_node();
           node->set_thread_lid( idx );