--- a/nestkernel/sparse_node_array.cpp
+++ b/nestkernel/sparse_node_array.cpp
@@ -26,6 +26,8 @@
 #include "exceptions.h"
 #include "node.h"
 
+#include "kernel_manager.h"
+
 nest::SparseNodeArray::NodeEntry::NodeEntry( Node& node, index node_id )
   : node_( &node )
   , node_id_( node_id )
@@ -38,7 +40,10 @@ nest::SparseNodeArray::SparseNodeArray()
   , max_node_id_( 0 )
   , local_min_node_id_( 0 )
   , local_max_node_id_( 0 )
-  , node_id_idx_scale_( 1.0 )
+  , scale_split_id_( 0 )
+  , scale_split_idx_( 0 )
+  , split_has_occured_( false )
+  , adding_nodes_with_proxies_( false ) // meaningless default
 {
 }
 
@@ -67,18 +72,40 @@ nest::SparseNodeArray::add_local_node( Node& node )
   if ( local_min_node_id_ == 0 ) // only first non-zero
   {
     local_min_node_id_ = node_id;
+    adding_nodes_with_proxies_ = node.has_proxies();
   }
   local_max_node_id_ = node_id;
   max_node_id_ = node_id;
 
-  // implies nodes_.size() > 1
-  if ( local_max_node_id_ > local_min_node_id_ )
+  if ( not split_has_occured_ )
   {
-    const double size = static_cast< double >( nodes_.size() - 1 );
-    node_id_idx_scale_ = size / ( local_max_node_id_ - local_min_node_id_ );
+    if ( adding_nodes_with_proxies_ != node.has_proxies() )
+    {
+      split_has_occured_ = true;
+    }
+    else
+    {
+      scale_split_id_ = node_id;
+      ++scale_split_idx_;
+      if ( local_max_node_id_ > local_min_node_id_ )
+      {
+        const double size = static_cast< double >( nodes_.size() - 1 );
+        id_idx_scale_[ 0 ] = size / ( local_max_node_id_ - local_min_node_id_ );
+      }
+    }
   }
-  assert( node_id_idx_scale_ > 0. );
-  assert( node_id_idx_scale_ <= 1. );
+  else
+  {
+    if ( local_max_node_id_ > scale_split_id_ )
+    {
+      const double size = static_cast< double >( nodes_.size() - scale_split_idx_ );
+      id_idx_scale_[ 1 ] = size / ( local_max_node_id_ - scale_split_id_ );
+    }
+  }
+  assert( id_idx_scale_[ 0 ] >= 0. );
+  assert( id_idx_scale_[ 0 ] <= 1. );
+  assert( id_idx_scale_[ 1 ] >= 0. );
+  assert( id_idx_scale_[ 1 ] <= 1. );
 }
 
 void
@@ -109,10 +136,14 @@ nest::SparseNodeArray::get_node_by_node_id( index node_id ) const
     return 0;
   }
 
+  const bool after_split = node_id > scale_split_id_;
+  const index ref_id = after_split ? scale_split_id_ : local_min_node_id_;
+  const size_t ref_idx = after_split ? scale_split_idx_ : 0;
+
   // now estimate index
-  const size_t idx_guess = std::min( static_cast< size_t >( std::floor( node_id_idx_scale_ * ( node_id - local_min_node_id_ ) ) ),
-  		                               nodes_.size() - 1 );
-  assert( idx_guess < nodes_.size() );
+  const size_t idx_guess =
+    std::min( static_cast< size_t >( ref_idx + std::floor( id_idx_scale_[ after_split ] * ( node_id - ref_id ) ) ),
+      nodes_.size() - 1 );
 
   // search left if necessary
   auto idx = idx_guess;
@@ -128,12 +159,12 @@ nest::SparseNodeArray::get_node_by_node_id( index node_id ) const
   }
 
   // adjust scaling based on search steps required
-  if ( std::abs( static_cast< long >( idx ) - static_cast< long >( idx_guess )) > 10 )
+  if ( std::abs( static_cast< long >( idx ) - static_cast< long >( idx_guess ) ) > 10 )
   {
-  	node_id_idx_scale_ = static_cast< double >( idx ) / ( node_id - local_min_node_id_ );
+    id_idx_scale_[ after_split ] = std::min( 1.0, static_cast< double >( idx - ref_idx ) / ( node_id - ref_id ) );
   }
-  assert( node_id_idx_scale_ >= 0.0 );
-  assert( node_id_idx_scale_ <= 1.0 );
+  assert( id_idx_scale_[ after_split ] >= 0.0 );
+  assert( id_idx_scale_[ after_split ] <= 1.0 );
 
   if ( idx < nodes_.size() and nodes_[ idx ].node_id_ == node_id )
   {