--- a/nestkernel/sparse_node_array.cpp
+++ b/nestkernel/sparse_node_array.cpp
@@ -24,15 +24,15 @@
 
 // Includes from nestkernel:
 #include "exceptions.h"
-#include "node.h"
-
 #include "kernel_manager.h"
+#include "node.h"
+#include "vp_manager_impl.h"
+
 
 nest::SparseNodeArray::NodeEntry::NodeEntry( Node& node, index node_id )
   : node_( &node )
   , node_id_( node_id )
 {
-  assert( node_id == node.get_node_id() );
 }
 
 nest::SparseNodeArray::SparseNodeArray()
@@ -40,11 +40,40 @@ nest::SparseNodeArray::SparseNodeArray()
   , max_node_id_( 0 )
   , local_min_node_id_( 0 )
   , local_max_node_id_( 0 )
-  , scale_split_id_( 0 )
-  , scale_split_idx_( 0 )
-  , split_has_occured_( false )
-  , adding_nodes_with_proxies_( false ) // meaningless default
+  , left_scale_( 1.0 )
+  , right_scale_( 1.0 )
+  , lookup_split_node_id_( 0 )
+  , lookup_split_idx_( 0 )
+  , have_split_( false )
+  , left_side_has_proxies_( false ) // meaningless initial value
+  , left_ctr_( 0 )
+  , right_ctr_( 0 )
 {
+}
+
+#include <iostream>
+nest::SparseNodeArray::~SparseNodeArray()
+{
+#pragma omp critical
+	{
+		std::cerr << left_ctr_ << '\t' << right_ctr_ << std::endl;
+
+	}
+}
+
+void
+nest::SparseNodeArray::clear()
+{
+  nodes_.clear();
+  max_node_id_ = 0;
+  local_min_node_id_ = 0;
+  local_max_node_id_ = 0;
+  left_scale_ = 1.0;
+  right_scale_ = 1.0;
+  lookup_split_node_id_ = 0;
+  lookup_split_idx_ = 0;
+  have_split_ = false;
+  left_side_has_proxies_ = false;
 }
 
 void
@@ -55,57 +84,44 @@ nest::SparseNodeArray::add_local_node( Node& node )
   // protect against node ID 0
   assert( node_id > 0 );
 
-  // local_min_node_id_ can only be 0 if no node has been stored
-  assert( local_min_node_id_ > 0 or nodes_.size() == 0 );
-
-  // local_min_node_id_ cannot be larger than local_max_node_id_
-  assert( local_min_node_id_ <= local_max_node_id_ );
-
-  // local_max_node_id_ cannot be larger than max_node_id_
-  assert( local_max_node_id_ <= max_node_id_ );
-
-  // node_id must exceed max_node_id_
+  // node_id must exceed max_node_id_ to ensure strictly increasing order
   assert( node_id > max_node_id_ );
 
   // all is consistent, register node and update auxiliary variables
   nodes_.push_back( NodeEntry( node, node_id ) );
-  if ( local_min_node_id_ == 0 ) // only first non-zero
-  {
-    local_min_node_id_ = node_id;
-    adding_nodes_with_proxies_ = node.has_proxies();
-  }
   local_max_node_id_ = node_id;
   max_node_id_ = node_id;
 
-  if ( not split_has_occured_ )
+  if ( local_min_node_id_ == 0 ) // We are adding the first node
   {
-    if ( adding_nodes_with_proxies_ != node.has_proxies() )
+    local_min_node_id_ = node_id;
+    left_side_has_proxies_ = node.has_proxies();
+
+    // We now know which scale applies on which side of the split
+    const double proxy_scale = 1.0 / static_cast< double >( kernel().vp_manager.get_num_virtual_processes() );
+    if ( left_side_has_proxies_ )
     {
-      split_has_occured_ = true;
+      left_scale_ = proxy_scale;
     }
     else
     {
-      scale_split_id_ = node_id;
-      ++scale_split_idx_;
-      if ( local_max_node_id_ > local_min_node_id_ )
-      {
-        const double size = static_cast< double >( nodes_.size() - 1 );
-        id_idx_scale_[ 0 ] = size / ( local_max_node_id_ - local_min_node_id_ );
-      }
+      right_scale_ = proxy_scale;
     }
   }
-  else
+
+  if ( not have_split_ )
   {
-    if ( local_max_node_id_ > scale_split_id_ )
+    if ( left_side_has_proxies_ != node.has_proxies() )
     {
-      const double size = static_cast< double >( nodes_.size() - scale_split_idx_ );
-      id_idx_scale_[ 1 ] = size / ( local_max_node_id_ - scale_split_id_ );
+      // node is first past splitting point
+      have_split_ = true;
+    }
+    else
+    {
+      lookup_split_node_id_ = node_id;  // update to last node so far on left side
+      ++lookup_split_idx_;              // index one beyond that node
     }
   }
-  assert( id_idx_scale_[ 0 ] >= 0. );
-  assert( id_idx_scale_[ 0 ] <= 1. );
-  assert( id_idx_scale_[ 1 ] >= 0. );
-  assert( id_idx_scale_[ 1 ] <= 1. );
 }
 
 void
@@ -119,12 +135,6 @@ nest::SparseNodeArray::update_max_node_id( index node_id )
 nest::Node*
 nest::SparseNodeArray::get_node_by_node_id( index node_id ) const
 {
-  // local_min_node_id_ cannot be larger than local_max_node_id_
-  assert( local_min_node_id_ <= local_max_node_id_ );
-
-  // local_max_node_id_ cannot be larger than max_node_id_
-  assert( local_max_node_id_ <= max_node_id_ );
-
   if ( node_id < 1 or max_node_id_ < node_id )
   {
     throw UnknownNode();
@@ -136,35 +146,33 @@ nest::SparseNodeArray::get_node_by_node_id( index node_id ) const
     return 0;
   }
 
-  const bool after_split = node_id > scale_split_id_;
-  const index ref_id = after_split ? scale_split_id_ : local_min_node_id_;
-  const size_t ref_idx = after_split ? scale_split_idx_ : 0;
+  /* Find base index and node ID for estimating location of desired node in array.
+   *
+   * The following lookup is safe at this point:
+   * - We have at least one local node, which by definition is on the left side, so nodes_[0] is fine.
+   * - left_side can only be false if we have at least one node on the right side, and then nodes_[lookup_split_idx_] is safe.
+   */
+  const bool left_side = node_id <= lookup_split_node_id_;
+  const double scale = left_side ? left_scale_ : right_scale_;
+  const size_t base_idx = left_side ? 0 : lookup_split_idx_;
+  const index base_id = nodes_[base_idx].node_id_;
 
-  // now estimate index
-  const size_t idx_guess =
-    std::min( static_cast< size_t >( ref_idx + std::floor( id_idx_scale_[ after_split ] * ( node_id - ref_id ) ) ),
-      nodes_.size() - 1 );
+  // estimate index, limit to array size for safety size
+  auto idx = std::min( static_cast< size_t >( base_idx + std::floor( scale * ( node_id - base_id ) ) ), nodes_.size() - 1 );
 
   // search left if necessary
-  auto idx = idx_guess;
   while ( 0 < idx and node_id < nodes_[ idx ].node_id_ )
   {
     --idx;
+    ++left_ctr_;
   }
 
   // search right if necessary
   while ( idx < nodes_.size() and nodes_[ idx ].node_id_ < node_id )
   {
     ++idx;
+    ++right_ctr_;
   }
-
-  // adjust scaling based on search steps required
-  if ( std::abs( static_cast< long >( idx ) - static_cast< long >( idx_guess ) ) > 10 )
-  {
-    id_idx_scale_[ after_split ] = std::min( 1.0, static_cast< double >( idx - ref_idx ) / ( node_id - ref_id ) );
-  }
-  assert( id_idx_scale_[ after_split ] >= 0.0 );
-  assert( id_idx_scale_[ after_split ] <= 1.0 );
 
   if ( idx < nodes_.size() and nodes_[ idx ].node_id_ == node_id )
   {