… nestkernel/target_identifier.h
… pynest/nest/lib/hl_api_nodes.py
… testsuite/pytests/test_getconnections.py
… testsuite/pytests/test_urbanczik_synapse.py