This PR revises the node-lookup heuristics of `SparseNodeArray` to provide fast lookup also in cases where many nodes without proxies (devices) are present. With this PR in place, good performance will be achieved if all devices are created either before or after all nodes with proxies.

The attached
[sra_demo.sli.gz](https://github.com/nest/nest-simulator/files/8030498/sra_demo.sli.gz) script illustrates this effect. Call it as
```
nest --userargs=num_sg,indegree,num_threads
```
The script creates 100.000 neurons and `num_sg` spike generators. It then connects neurons to neurons and generators to neurons with the given fixed `indegree`, multapses permitted, running on `num_threads`. The interesting point is to see performance differences for small (eg 1) and large (eg 25.000) `num_sg` for serial and thread-parallel cases. Current master (cc15d0590) performs very poorly on multiple threads, much worse than serial, because the linear lookup heuristics becomes practically useless when nodes distributed across processes are mixed with a large number of non-distributed devices. This PR (split-sparse-array@a9cc70dce) notes where in the SparseNodeArray the boundary between normal nodes and devices is and uses separate linear lookup in each part. To keep complexity within bounds, it only allows a single boundary, so neurons and devices should be created en bloc, not intermingled.

Benchmark results for indegree==100 from my laptop (macOS 12.2, 2 GHz Intel Core i5, 4 cores, Clang 13), where the rightmost column is most important:
 
|  | | 1 SG, 1 Thread | 1 SG 4 Threads | 25k SGs, 1 Thread | 25k SGs, 4 Threads |
|--|--|--:|--:|--:|--:|
|Master| N->N| 2.1s | 0.4s | 2.1s |18.3s|
|           | G->N| 0.8s |  0.3s | 4.7s|21.4s|
|This PR|N->N|2.6s |0.4s |2.1s |0.5s |
|            |G->N| 1.1s| 0.3s| 4.7s|2.1s | 


To do:
- Review wether to keep adjustment of scale based on lookup experience
- Review field names
- Add user documentation pointing out the need to not intermingle creation of neurons and devices