@heplesser 2022-02-16
@stinebuu Thanks for the review, all should be fixed now.

BTW, I created a branch https://github.com/heplesser/nest-simulator/tree/split-sparse-array-with-tracing with a version that prints additional tracing info (steps taken in search) when NEST terminates. That is not to be merged, just for exploration/debugging.