@mlober 2022-03-07
I tested the example script and can confirm the acceleration of this PR compared to the master version.